requirements.txt
streamlit
pandas
supabase
python-dateutil
requests